put future-related methods to AbstractManagedExecutorService
Move them from ManagedExecutorServiceImpl and 
ManagedScheduledExecutorServiceImpl as they will be implemented in the
same way.
